Want simple performance tool
This tool allows us to pump IO full speed to supported storage URI's.
The intent of this is to be able to test performance without the need
for setting up any targets and, subsequently, benchmark tools.

Its no where near as sophisticated as fio, but thats also not the
intent. The intent is to get good numbers but also, to assist in
testing/developing the handling of IO errors and rebuild

+++ mayastor/src/rebuild/rebuild_impl.rs
@@ -156,7 +156,7 @@ impl RebuildJob {
 
         for _ in 0 .. tasks.total {
             let copy_buffer = destination_hdl
-                .dma_malloc((segment_size_blks * block_size) as usize)
+                .dma_malloc(segment_size_blks * block_size)
                 .context(NoCopyBuffer {})?;
             tasks.tasks.push(RebuildTask {
                 buffer: copy_buffer,
@@ -323,7 +323,7 @@ impl RebuildJob {
 
             copy_buffer = self
                 .destination_hdl
-                .dma_malloc((segment_size_blks * self.block_size) as usize)
+                .dma_malloc(segment_size_blks * self.block_size)
                 .context(NoCopyBuffer {})?;
 
             &mut copy_buffer
